requirements.txt
mayavi==4.8.2
numpy==2.1.2
Pillow==11.0.0
PyQt5==5.15.11
PyQt5_sip==12.15.0
